+++ examples/Backtesting/run_hybrid.py
@@ -15,19 +15,17 @@ import matplotlib.pyplot as plt
 import numpy as np
 import seaborn as sns
 
-from baybe.core import BayBE
-
-from baybe.parameters import NumericContinuous, NumericDiscrete
+from baybe import BayBE
+from baybe.parameters import NumericalContinuousParameter, NumericalDiscreteParameter
 from baybe.searchspace import SearchSpace
 from baybe.simulation import simulate_scenarios
-from baybe.strategies.bayesian import (
+from baybe.strategies import (
     NaiveHybridRecommender,
+    RandomRecommender,
     SequentialGreedyRecommender,
+    Strategy,
 )
-from baybe.strategies.sampling import RandomRecommender
-from baybe.strategies.strategy import Strategy
 from baybe.targets import NumericalTarget, Objective
-
 
 # For the full simulation, we need to define some additional parameters.
 # These are the number of Monte Carlo runs and the number of experiments to be
@@ -83,18 +81,18 @@ if set(CONT_INDICES + DISC_INDICES) != set(range(DIMENSION)):
         "indices do not match."
     )
 
-# Construct the continuous parameters as NumericContinuous parameters
+# Construct the continuous parameters
 cont_parameters = [
-    NumericContinuous(
+    NumericalContinuousParameter(
         name=f"x_{k+1}",
         bounds=(BOUNDS[k]),
     )
     for k in CONT_INDICES
 ]
 
-# Construct the discrete parameters as NumericDiscrete parameters
+# Construct the discrete parameters
 disc_parameters = [
-    NumericDiscrete(
+    NumericalDiscreteParameter(
         name=f"x_{k+1}",
         values=list(np.linspace(*BOUNDS[k], POINTS_PER_DIM)),
         tolerance=0.01,