+++ examples/Searchspaces/continuous_space_botorch_function.py
@@ -3,12 +3,11 @@ Example for using synthetic test functions in continuous spaces using a wrapped 
 function as synthetic test function.
 """
 
-from baybe.core import BayBE
-from baybe.parameters import NumericContinuous
+from baybe import BayBE
+from baybe.parameters import NumericalContinuousParameter
 from baybe.searchspace import SearchSpace
 from baybe.targets import NumericalTarget, Objective
-
-from baybe.utils.botorch_wrapper import botorch_function_wrapper
+from baybe.utils import botorch_function_wrapper
 
 # Import the desired test function from botorch here
 from botorch.test_functions import Rastrigin
@@ -45,10 +44,10 @@ BOUNDS = TestFunction.bounds
 # Create the wrapped function itself.
 WRAPPED_FUNCTION = botorch_function_wrapper(test_function=TestFunction)
 
-# Since this is the continuous test, we only construct NumericContinuous parameters.
+# Since this is the continuous test, we only construct numerical continuous parameters.
 # We use that data of the test function to deduce bounds and number of parameters.
 parameters = [
-    NumericContinuous(
+    NumericalContinuousParameter(
         name=f"x_{k+1}",
         bounds=(BOUNDS[0, k], BOUNDS[1, k]),
     )