+++ examples/Searchspaces/hybrid_space.py
@@ -9,14 +9,12 @@ continuous_space example.
 
 import numpy as np
 
-from baybe.core import BayBE
-from baybe.parameters import NumericContinuous, NumericDiscrete
+from baybe import BayBE
+from baybe.parameters import NumericalContinuousParameter, NumericalDiscreteParameter
 from baybe.searchspace import SearchSpace
-from baybe.strategies.bayesian import NaiveHybridRecommender
-from baybe.strategies.strategy import Strategy
+from baybe.strategies import NaiveHybridRecommender, Strategy
 from baybe.targets import NumericalTarget, Objective
-
-from baybe.utils.botorch_wrapper import botorch_function_wrapper
+from baybe.utils import botorch_function_wrapper
 
 # Import the desired test function from botorch here
 from botorch.test_functions import Rastrigin
@@ -76,18 +74,18 @@ if set(CONT_INDICES + DISC_INDICES) != set(range(DIMENSION)):
         "indices do not match."
     )
 
-# Construct the continuous parameters as NumericContinuous parameters
+# Construct the continuous parameters
 cont_parameters = [
-    NumericContinuous(
+    NumericalContinuousParameter(
         name=f"x_{k+1}",
         bounds=(BOUNDS[0, k], BOUNDS[1, k]),
     )
     for k in CONT_INDICES
 ]
 
-# Construct the discrete parameters as NumericDiscrete parameters
+# Construct the discrete parameters
 disc_parameters = [
-    NumericDiscrete(
+    NumericalDiscreteParameter(
         name=f"x_{k+1}",
         values=list(np.linspace(BOUNDS[0, k], BOUNDS[1, k], POINTS_PER_DIM)),
         tolerance=0.01,