+++ examples/Serialization/create_from_config.py
@@ -8,14 +8,14 @@ actual class instances you can create such a config by providing a dictionary wi
 "type":"name of the class". Example: FPSRecommender() becomes {"type": "FPSRecommender"}
 """
 
-from baybe.core import BayBE
+from baybe import BayBE
 
 
 CONFIG = """
 {
     "parameters": [
         {
-            "type": "Categorical",
+            "type": "CategoricalParameter",
             "name": "Granularity",
             "values": [
                 "coarse",
@@ -25,7 +25,7 @@ CONFIG = """
             "encoding": "OHE"
         },
         {
-            "type": "NumericDiscrete",
+            "type": "NumericalDiscreteParameter",
             "name": "Pressure[bar]",
             "values": [
                 1,
@@ -35,7 +35,7 @@ CONFIG = """
             "tolerance": 0.2
         },
         {
-            "type": "GenericSubstance",
+            "type": "SubstanceParameter",
             "name": "Solvent",
             "data": {
                 "Solvent A": "COC",