+++ tests/test_input_output.py
@@ -4,8 +4,7 @@ Tests for basic input-output nad iterative loop.
 import numpy as np
 import pytest
 
-from baybe.utils import add_fake_results
-
+from baybe.utils.dataframe import add_fake_results
 
 # List of tests that are expected to fail (still missing implementation etc)
 param_xfails = []