+++ tests/test_iterations.py
@@ -7,16 +7,14 @@ import pytest
 
 from baybe.searchspace import SearchSpaceType
 from baybe.strategies.bayesian import (
+    BayesianRecommender,
     NaiveHybridRecommender,
     SequentialGreedyRecommender,
 )
-from baybe.strategies.recommender import (
-    BayesianRecommender,
-    NonPredictiveRecommender,
-    Recommender,
-)
+from baybe.strategies.recommender import NonPredictiveRecommender, Recommender
 from baybe.surrogate import SurrogateModel
-from baybe.utils import add_fake_results, add_parameter_noise, get_subclasses
+from baybe.utils.basic import get_subclasses
+from baybe.utils.dataframe import add_fake_results, add_parameter_noise
 
 ########################################################################################
 # Settings of the individual components to be tested